# OpenAPI 3.0.json
openapi: 3.0.3
info:
  title: Multi-tenant Commerce Ingestion API
  version: 1.0.0
  description: >
    CRUD + search APIs for Organizations, Websites, Orders, Fulfillments, and Tracking.
    Supports date-time range filtering and pagination (page/size).
servers:
  - url: https://api.example.com
...
paths:
  /organizations:
    post:
      tags: [Organizations]
      summary: Create organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Organizations]
      summary: List/search organizations (date range + pagination)
      operationId: listOrganizations
      parameters:
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrgStatus"
        - name: name
          in: query
          description: Name contains (case-insensitive, implementation-specific)
          schema:
            type: string
      responses:
        "200":
          description: Paged organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedOrganization"
        "400":
          $ref: "#/components/responses/BadRequest"
  /organizations/search:
    get:
      tags: [Organizations]
      summary: Search organizations by external/reference id
      operationId: searchOrganizationsByRef
      parameters:
        - name: externalId
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
      responses:
        "200":
          description: Paged organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedOrganization"
        "400":
          $ref: "#/components/responses/BadRequest"
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization by id
      operationId: getOrganizationById
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Organizations]
      summary: Update organization (full replace)
      operationId: updateOrganization
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Organizations]
      summary: Update organization (partial)
      operationId: patchOrganization
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Organizations]
      summary: Delete organization
      operationId: deleteOrganization
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{orgId}/websites:
    post:
      tags: [Websites]
      summary: Create website for organization
      operationId: createWebsite
      parameters:
        - $ref: "#/components/parameters/orgIdPathAlt"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Websites]
      summary: List/search websites for organization (date range + pagination)
      operationId: listWebsites
      parameters:
        - $ref: "#/components/parameters/orgIdPathAlt"
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/WebsiteStatus"
        - name: platform
          in: query
          schema:
            $ref: "#/components/schemas/Platform"
        - name: code
          in: query
          schema:
            type: string
        - name: domain
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paged websites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedWebsite"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{orgId}/websites/search:
    get:
      tags: [Websites]
      summary: Search websites within an organization by websiteId/code/domain
      operationId: searchWebsites
      parameters:
        - $ref: "#/components/parameters/orgIdPathAlt"
        - name: websiteId
          in: query
          description: Website UUID (returns one item if exact match)
          schema:
            type: string
            format: uuid
        - name: code
          in: query
          schema:
            type: string
        - name: domain
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
      responses:
        "200":
          description: Paged websites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedWebsite"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{orgId}/websites/{websiteId}:
    get:
      tags: [Websites]
      summary: Get website by id
      operationId: getWebsiteById
      parameters:
        - $ref: "#/components/parameters/orgIdPathAlt"
        - $ref: "#/components/parameters/websiteIdPath"
      responses:
        "200":
          description: Website
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Websites]
      summary: Update website (full replace)
      operationId: updateWebsite
      parameters:
        - $ref: "#/components/parameters/orgIdPathAlt"
        - $ref: "#/components/parameters/websiteIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Websites]
      summary: Update website (partial)
      operationId: patchWebsite
      parameters:
        - $ref: "#/components/parameters/orgIdPathAlt"
        - $ref: "#/components/parameters/websiteIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsitePatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Websites]
      summary: Delete website
      operationId: deleteWebsite
      parameters:
        - $ref: "#/components/parameters/orgIdPathAlt"
        - $ref: "#/components/parameters/websiteIdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    from:
      name: from
      in: query
      required: false
      description: Start datetime (inclusive), ISO-8601
      schema:
        type: string
        format: date-time
    to:
      name: to
      in: query
      required: false
      description: End datetime (inclusive), ISO-8601
      schema:
        type: string
        format: date-time
    page:
      name: page
      in: query
      required: false
      description: 0-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      required: false
      description: Batch size / page size
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    sort:
      name: sort
      in: query
      required: false
      description: Sorting, e.g. createdAt,desc or updatedAt,asc
      schema:
        type: string
        default: updatedAt,desc
    orgIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    orgIdPathAlt:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    websiteIdPath:
      name: websiteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required: [timestamp, status, error, message, path]

    # ----------------------------
    # Enums
    # ----------------------------
    OrgStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    WebsiteStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    Platform:
      type: string
      enum: [SHOPIFY, NETSUITE, CUSTOM, MAGENTO, OTHER]
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          nullable: true
        name:
          type: string
        status:
          $ref: "#/components/schemas/OrgStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, status, createdAt, updatedAt]
    OrganizationCreateRequest:
      type: object
      properties:
        externalId:
          type: string
          nullable: true
        name:
          type: string
          minLength: 2
        status:
          $ref: "#/components/schemas/OrgStatus"
      required: [name]
    OrganizationUpdateRequest:
      type: object
      properties:
        externalId:
          type: string
          nullable: true
        name:
          type: string
          minLength: 2
        status:
          $ref: "#/components/schemas/OrgStatus"
      required: [name, status]
    OrganizationPatchRequest:
      type: object
      properties:
        externalId:
          type: string
          nullable: true
        name:
          type: string
          minLength: 2
        status:
          $ref: "#/components/schemas/OrgStatus"
    PagedOrganization:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
      required: [data, page, size, totalElements, totalPages, hasNext]

    # ----------------------------
    # Website
    # ----------------------------
    Website:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        platform:
          $ref: "#/components/schemas/Platform"
        domain:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/WebsiteStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, orgId, code, name, platform, status, createdAt, updatedAt]
    WebsiteCreateRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 2
        name:
          type: string
          minLength: 2
        platform:
          $ref: "#/components/schemas/Platform"
        domain:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/WebsiteStatus"
      required: [code, name, platform]
    WebsiteUpdateRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 2
        name:
          type: string
          minLength: 2
        platform:
          $ref: "#/components/schemas/Platform"
        domain:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/WebsiteStatus"
      required: [code, name, platform, status]
    WebsitePatchRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 2
        name:
          type: string
          minLength: 2
        platform:
          $ref: "#/components/schemas/Platform"
        domain:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/WebsiteStatus"
    PagedWebsite:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Website"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
      required: [data, page, size, totalElements, totalPages, hasNext]

    # ----------------------------
    # Order
    # ----------------------------
...